Move Reminder model to rmndin.reminders.models
